refactor(util): remove dead cache util, relocate pure leaves out of util/
Deletes util/cacheUtils.ts (zero importers, superseded by util/simpleCache.ts).
Moves util/models.ts to types/models.ts (pure interfaces, no runtime logic)
and util/syncMeeting.ts to services/syncMeeting.ts (thin fetch wrapper,
belongs with the other service wrappers). Groups the remaining
domain-agnostic helpers (breakpoints, linkify, simpleCache, color) under
util/common/. Pure relocation, no logic changes.
