+++ frontend/app/api/export/lease/route.ts
@@ -2,7 +2,7 @@ import { Role } from "@prisma/client";
 import { requireRole } from "../../../../services/auth";
 import { defaultLeaseSettings } from "../../../../util/leaseDefaults";
 import { formatDayColumn } from "../../../../util/recurrenceDisplay";
-import type { ILeaseSettings, IRoomRate } from "../../../../util/models";
+import type { ILeaseSettings, IRoomRate } from "../../../../types/models";
 import { prisma } from "../../../../lib/prisma";
 
 const notDeleted = { deletedAt: null };
