+++ frontend/app/components/admin/export/ExportTab.tsx
@@ -11,7 +11,7 @@ import {
   SIGNAGE_ZOOM_SLUGS,
 } from "../../../../util/signageFilters";
 import { ROOM_COLORS, ZOOM_ROOM_COLOR, CATEGORY_COLOR } from "../../../../util/filterColors";
-import type { ILeaseSettings } from "../../../../util/models";
+import type { ILeaseSettings } from "../../../../types/models";
 import { ALL_MEETING_EXPORT_FIELD_KEYS, type MeetingExportFieldKey } from "../../../../util/meetingExportFields";
 import type { LeaseYearCycle } from "../../../../util/leaseYearCycles";
 import FilterGroup, { FilterGroupItem } from "../../shared/FilterGroup";
