+++ frontend/app/components/meeting-form/EditMeeting.tsx
@@ -13,7 +13,7 @@ import ConflictOverrideModal from './ConflictOverrideModal';
 import CloseIcon from '@mui/icons-material/Close';
 import IconButton from '@mui/material/IconButton';
 
-import { IMeeting } from '../../../util/models'
+import { IMeeting } from '../../../types/models'
 import { physicalRoomOptions, zoomRoomOptions } from "../../../util/rooms";
 import { useMeetingForm, CAL_TYPE_OPTIONS, CAL_TYPE_COLOR, DESCRIPTION_MAX_LENGTH } from '../../../hooks/useMeetingForm';
 import { ConflictListRow } from '../../../util/conflictDisplay';
