+++ frontend/app/components/meeting-form/ViewMeeting.tsx
@@ -9,15 +9,15 @@ import ResumeMeetingModal from './ResumeMeetingModal';
 import TagList from '../atoms/TagList';
 import BottomSheet from '../atoms/BottomSheet';
 
-import { IRecurrencePattern } from '../../../util/models';
+import { IRecurrencePattern } from '../../../types/models';
 import { formatCompactTimeRange, formatMeetingDateLine } from "../../../util/timeFormat";
 import { formatETDateString } from "../../../util/timeUtils";
-import { retryMeetingSync } from "../../../util/syncMeeting";
+import { retryMeetingSync } from "../../../services/syncMeeting";
 import { formatSuspensionStatusText } from "../../../util/suspensionText";
 import { ROOM_COLORS, ZOOM_ROOM_COLOR } from "../../../util/filterColors";
 import { formatRecurrencePattern } from "../../../util/recurrenceDisplay";
 import { isZoomRoomMismatched } from "../../../util/rooms";
-import { linkify } from "../../../util/linkify";
+import { linkify } from "../../../util/common/linkify";
 import { zoomHostLabel } from "../../../util/zoomHosts";
 import { MODE_ICON_SRC } from "../../../util/modeIcons";
 import { useZoomHostPool } from "../../../hooks/useZoomHostPool";
