+++ frontend/app/components/meeting-form/ZoomHostField.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useRef, useState } from 'react';
 import Dropdown from '../atoms/Dropdown';
 import { useZoomHostPool } from '../../../hooks/useZoomHostPool';
 import { zoomHostLabel } from '../../../util/zoomHosts';
-import { IMeeting } from '../../../util/models';
+import { IMeeting } from '../../../types/models';
 import styles from '../../../styles/components/meeting-form/ZoomHostField.module.scss';
 
 const AUTOMATIC = 'Automatic assignment';
