+++ frontend/tests/integration/update-meeting-route.test.ts
@@ -1,5 +1,5 @@
 import { randomUUID } from "crypto";
-import type { IMeeting } from "../../util/models";
+import type { IMeeting } from "../../types/models";
 
 // Next's after() throws when called outside a real request scope, which route handlers
 // invoked directly (not through the Next server) always are. The sync promise passed to
