+++ frontend/util/common/simpleCache.ts
