+++ frontend/util/suspension.ts
@@ -3,7 +3,7 @@ import { Meeting, RecurrencePattern, SuspensionPeriod } from "@prisma/client";
 import { formatETDateString } from "./timeUtils";
 import { isDateSuspended, adjustOccurrenceToDate, firstOccurrenceOnOrAfter } from "./meetingOccurrences";
 import { calendarIdsForMeeting, createCalendarEvent, deleteCalendarEvent } from "../services/googleCalendar";
-import { IMeeting } from "./models";
+import { IMeeting } from "../types/models";
 import { prisma } from "../lib/prisma";
 
 export type MeetingWithPattern = Meeting & { recurrencePattern: RecurrencePattern | null };
